Adding Exception handling to controllers

+++ src/main/java/com/gabo/springdata/shoppingcar/controllers/UserController.java
@@ -7,9 +7,11 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import com.gabo.springdata.shoppingcar.entities.User;
+import com.gabo.springdata.shoppingcar.exceptions.NotFoundException;
 import com.gabo.springdata.shoppingcar.repos.UserRepository;
 
 import jakarta.transaction.Transactional;
+import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("/Users")
@@ -21,37 +23,48 @@ public class UserController {
 		this.repository = repository;
 	}
 	
-	///////Insert user into table
+
+	//Insert user into table
 	@PostMapping("/insert")
-	public User saveUser(@Validated @RequestBody User user) {
-		List<User> tempUser = repository.findByEmail(user.getName());
-		if(tempUser == null || tempUser.size() == 0) {
+	public User insertUser(@Valid @RequestBody User user) {
+		List<User> tempUser = repository.findByEmail(user.getName()); 
+		if(tempUser.isEmpty()) {
 			return repository.save(user);
 		}
 		return user;
 	}
 	
-	///////Update an existing user
+	
+	//Update an existing user
 	@PostMapping("/update/{id}")
 	public User updateUserById( @PathVariable("id") Integer id,@Validated @RequestBody User user) {
 		User tempUser = repository.findById(id).get();
-		tempUser.setArea_of_interest(user.getArea_of_interest());
-		tempUser.setEmail(user.getEmail());
-		return repository.save(tempUser);
+		if(tempUser == null) {
+			throw new NotFoundException("No match for id: "+id);
+		}else {
+			tempUser.setAreaOfInterest(user.getAreaOfInterest());
+			tempUser.setEmail(user.getEmail());
+			return repository.save(tempUser);
+		}
+	
 	}
 	
 	
-	///////Delete user
+	//Delete user
 	@Transactional
 	@PostMapping("/delete/{id}")
 	public User deleteUserById(@PathVariable("id") Integer id) {
 		User user = repository.findById(id).get();
-		repository.deleteById(id);
-		return user;
+		if(user == null) {
+			throw new NotFoundException("No match for id: "+id);
+		}else {
+			repository.deleteById(id);
+			return user;
+		}
+		
 	}
 	
-	
-	////////Retrieve list of users
+	//Retrieve list of users
 	@GetMapping("/showUsers")
 	public List<String> getUsers(){
 		List<String> names = new ArrayList<String>();
@@ -63,22 +76,27 @@ public class UserController {
 	}
 	
 	
-	////////Retrieve specific user filtered by name
+	//Retrieve specific user filtered by name
 	@GetMapping("/showUsers/name/{name}")
 	public List<User> getUsersByName(@PathVariable("name") String name) {
-		return repository.findByName(name);
+		List<User> user = repository.findByName(name);
+		if(user.isEmpty()) {
+			throw new NotFoundException("No user named: "+ name);
+		}else {
+			return user;
+		}
 	}
 	
 	
-	////////Retrieve specific user filtered by name
+	//Retrieve specific user filtered by email
 	@GetMapping("/showUsers/email/{email}")
 	public List<User> getUsersByEmail(@PathVariable("email") String email) {
-		return repository.findByEmail(email);
+		List<User> user = repository.findByEmail(email);
+		if(user.isEmpty()) {
+			throw new NotFoundException("No user with email: "+ email);
+		}else {
+			return user;
+		}
 	}
-	
-	
-	
-	
-	
 	
 }
